Improved search for Loco properties
Revised the way that the Loco properties list is parsed. 
IF a loco with the same name has been seen before, code now assumes it has read the whole list.
This has TWO important changes in how the throttle works: 
  IF Rocrail (or any other throttle) commands a speed change before the throttle has read the whole Loco table, the throttle now becomes a "single Loco" throttle, and can ONLY control that loco. 
 BUT IF the select button on the throttle is pressed before the throttle sees any speed commands, then it will read the whole Rocrail loco list and can select ANY loco to control.

ALSO, pressing Buttonright to scroll through the menu levels now does not reset speed to zero, unless you are selecting a new loco to control.